Adding markers for poll locations and home.
