+++ src/components/Geocoder/Geocoder.js
@@ -11,6 +11,21 @@ function Geocoder(props) {
     ...(props.state || {})
   };
 
+  const geocoderAnimation = () => {
+    console.log('start animation');
+    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. '; }, 5000);
+    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 4'; }, 5500);
+    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 40'; }, 6000);
+    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 400'; }, 6500);
+    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 4000'; }, 7000);
+    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 4000 v'; }, 7500);
+    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 4000 ve'; }, 8000);
+    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 4000 ver'; }, 8500);
+    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 4000 vern'; }, 9000);
+    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 4000 verno'; }, 9500);
+    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 4000 vernor'; }, 10000);
+  }
+
   const getAddressSuggestions = (addr) => {
     let tempAddr = addr.split(",");
     tempAddr = tempAddr[0];
