+++ src/components/Map/Map.js
@@ -1,5 +1,5 @@
 import React, { useState }from 'react';
-import ReactMapboxGl from "react-mapbox-gl";
+import ReactMapboxGl, { Marker } from "react-mapbox-gl";
 import './Map.scss';
 const MapGL = ReactMapboxGl({
   accessToken:
@@ -14,28 +14,24 @@ function Map(props) {
     ...(props.state || {})
   };
 
-  const geocoderAnimation = () => {
-    console.log('start animation');
-    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. '; }, 5000);
-    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 4'; }, 5500);
-    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 40'; }, 6000);
-    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 400'; }, 6500);
-    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 4000'; }, 7000);
-    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 4000 v'; }, 7500);
-    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 4000 ve'; }, 8000);
-    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 4000 ver'; }, 8500);
-    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 4000 vern'; }, 9000);
-    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 4000 verno'; }, 9500);
-    setTimeout(()=>{ document.querySelector('#geocoder input').placeholder = 'Ex. 4000 vernor'; }, 10000);
-  }
-
   return (
     <article id="Map">
       <MapGL 
       style="mapbox://styles/mapbox/streets-v9"
       zoom={[13]}
       center={[-83.1, 42.36]}
-      ></MapGL>
+      >
+        <Marker
+          id="home"
+          coordinates={[-83.1, 42.36]}>
+          <img  src="https://via.placeholder.com/30"/>
+        </Marker>
+        <Marker
+          id="poll"
+          coordinates={[-83.1, 42.37]}>
+          <img  src="https://via.placeholder.com/30"/>
+        </Marker>
+      </MapGL>
     </article>
   );
 }
